add Status filter + only show per_page if rows
