+++ autoreduce_frontend/templates/search.html
@@ -1,6 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Search{% endblock %}
 {% load static %}
+{% include crispy_tags %}
 {% load crispy_forms_tags %}
 {% load render_table from django_tables2 %}
 
@@ -8,34 +9,30 @@
     <h1>Search Page</h1>
 
     <form action="" method="get">
-        {{filter.form.run_number|as_crispy_field:"bootstrap"}}
-        <br>
-        {{filter.form.instrument|as_crispy_field:"bootstrap"}}
-        <br>
-        {{filter.form.run_description|as_crispy_field:"bootstrap"}}
-        <input type="radio" id="contains" name="run_description_qualifier" value="contains" {% if run_description_qualifier == 'contains' %} checked {% endif %}>
-        <label for="contains">Contains</label>
-        <input type="radio" id="exact" name="run_description_qualifier" value="exact" {% if run_description_qualifier == 'exact' %} checked {% endif %}>
-        <label for="exact">Exact</label>
-        <br>
-        {{filter.form.created|as_crispy_field:"bootstrap"}}
-        <br>
+        {{filter.form.run_number|as_crispy_field}}
+
+        {{filter.form.instrument|as_crispy_field}}
+
+        {{filter.form.run_description|as_crispy_field}}
+        <div class="radios">
+            <input type="radio" id="contains" name="run_description_qualifier" value="contains" {% if run_description_qualifier == 'contains' %} checked {% endif %}>
+            <label for="contains">Contains</label>
+            <input type="radio" id="exact" name="run_description_qualifier" value="exact" {% if run_description_qualifier == 'exact' %} checked {% endif %}>
+            <label for="exact">Exact</label>
+        </div>
+
+        {{filter.form.created|as_crispy_field}}
+
+        {{filter.form.status|as_crispy_field}}
+
         <button id="search-button" type="submit" class="btn btn-primary">Search</button>
     </form>
-    <br>
+    <hr/>
 
     <div>
-        <label for="pagination_select">Items per page</label>
-            <select title="The number of reduction jobs that should be shown per page" class="form-control" id="select_per_page" name="per_page">
-                <option {% if per_page == '10' %} selected="selected" {% endif %}>10</option>
-                <option {% if per_page == '25' %} selected="selected" {% endif %}>25</option>
-                <option {% if per_page == '50' %} selected="selected" {% endif %}>50</option>
-                <option {% if per_page == '100' %} selected="selected" {% endif %}>100</option>
-                <option {% if per_page == '250' %} selected="selected" {% endif %}>250</option>
-                <option {% if per_page == '500' %} selected="selected" {% endif %}>500</option>
-            </select>
-        <br>
-        <button class="btn btn-primary" id="apply_filters" onclick="update_page()">Apply Filters</button>
+        {% if table.rows %}
+            {{options_form.pagination|as_crispy_field}} 
+        {% endif %}
     </div>
     
     <br>
